extract cell and apply pointer to remove button
