+++ Core/VariantManager.swift
@@ -47,8 +47,8 @@ public struct Variant {
         Variant(name: "sd", weight: doNotAllocate, features: []),
         Variant(name: "se", weight: doNotAllocate, features: []),
 
-        // iPad improvements
-        Variant(name: "mc", weight: 1, features: []),
+        // iPad improvements (WIP)
+        // Variant(name: "mc", weight: 1, features: []),
         Variant(name: "md", weight: 1, features: [.iPadImprovements])
 
     ]
