+++ DuckDuckGo.xcodeproj/xcshareddata/xcschemes/DuckDuckGo.xcscheme
@@ -108,7 +108,7 @@
          <EnvironmentVariable
             key = "VARIANT"
             value = "md"
-            isEnabled = "YES">
+            isEnabled = "NO">
          </EnvironmentVariable>
          <EnvironmentVariable
             key = "ONBOARDING"
